LAB_01_Chatbot
A chatbot consisting of the predefined dataset, which is capable of having a casual conversation with human. It consists of a dataset with specific responses for a specific type of conversations done by the user. It learns from the sentence tags, defining each sentence with a respective
tag. It uses bag of words model from NLP text modelling technique