Lab_02_Implementation of Fuzzy Logic
Fuzzy Logic is a method of reasoning that resembles human
reasoning. The approach of Fuzzy logic imitates the way of decision making in humans that involves all intermediate possibilities between digital values YES and NO. The fuzzy logic works on the levels of possibilities of input to achieve the definite output.